updated rcas.Rmd to make each chunk optional to run
According to the input parameters passed to the rcas.Rmd, the reporting script conditionally evaluates each chunk.
Table and figure counts, table and figure captions, section headers are all conditionally printed to the report file.
To turn off a chunk related to a module (e.g. GO term calculation for biological processes) the related input argument should be passed as 'NOT_RUN'
instead of passing a path to the file containing the related data to be plotted in the report file.